Refactor: generic versus YouTube tests
