+++ firefox_media_tests/playback/manifest.ini
@@ -1,2 +1,1 @@
-[test_prefs.py]
-[test_basic_playback.py ]
+[test_video_playback.py]
