--- firefox_media_tests/playback/test_basic_playback.py
@@ -1,37 +1,0 @@
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-from media_test_harness.testcase import MediaTestCase
-from marionette_driver import Wait
-
-from firefox_media_tests.utils import (playback_done, playback_started,
-                                       verbose_until, wait_for_almost_done)
-from media_utils.media_player import YouTubePuppeteer
-
-
-class TestVideoPlayback(MediaTestCase):
-
-    def setUp(self):
-        MediaTestCase.setUp(self)
-        self.test_urls = self.video_urls
-
-    def tearDown(self):
-        MediaTestCase.tearDown(self)
-
-    def test_video_playing_in_one_tab(self):
-        with self.marionette.using_context('content'):
-            for url in self.test_urls:
-                youtube = YouTubePuppeteer(self.marionette, url)
-                youtube.deactivate_autoplay()
-                wait_for_almost_done(youtube, final_piece=60)
-                verbose_until(Wait(youtube, timeout=300), youtube,
-                              playback_done)
-
-    def test_playback_starts(self):
-        with self.marionette.using_context('content'):
-            for url in self.test_urls:
-                youtube = YouTubePuppeteer(self.marionette, url)
-                # TODO getting timeout on nightly sometimes.
-                verbose_until(Wait(youtube, timeout=60), youtube,
-                              playback_started)
