+++ firefox_media_tests/playback/test_video_playback.py
@@ -3,17 +3,21 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 
-from marionette_driver.errors import TimeoutException
 from marionette_driver import Wait
+from marionette_driver.errors import TimeoutException
 
-from firefox_media_tests.utils import (verbose_until)
+from firefox_media_tests.utils import verbose_until
 from media_test_harness.testcase import MediaTestCase
 from media_utils.video_puppeteer import (playback_done, playback_started,
                                          VideoPuppeteer, VideoException)
 
 
 class TestVideoPlaybackBase(MediaTestCase):
+    """ Test MSE playback in HTML5 video element.
 
+    These tests should pass on any site where a single video element plays
+    upon loading and is uninterrupted (by ads, for example)
+    """
     def setUp(self):
         MediaTestCase.setUp(self)
         self.test_urls = self.video_urls
@@ -28,6 +32,10 @@ class TestVideoPlaybackBase(MediaTestCase):
                 verbose_until(Wait(video, timeout=30), video,
                               playback_started)
                 video.pause()
+                src = video.video_src
+                if not src.startswith('mediasource'):
+                    self.marionette.log('video is not mediasource: %s' % src,
+                                        level='WARNING')
 
     def test_video_playback_partial(self):
         self.run_playback(timeout=120)
@@ -39,19 +47,29 @@ class TestVideoPlaybackBase(MediaTestCase):
         with self.marionette.using_context('content'):
             for url in self.test_urls:
                 video = VideoPuppeteer(self.marionette, url)
+                # Some vendors don't start the videos at the beginning,
+                # remembering where you were last time you watched.
+                start_time = video.current_time
                 verbose_until(Wait(video, timeout=30), video,
                               playback_started)
                 duration_timeout = video.duration * 1.3
+                time_delta = 2
                 if timeout <= 0:
                     duration = duration_timeout
                 else:
+                    # Looser constraint for partial playback
+                    time_delta = 5
                     duration = min(timeout, duration_timeout)
 
                 try:
-                    # Some vendors don't start the videos at the beginning,
-                    # remembering where you were last time you watched.
-                    start_time = video.current_time
                     verbose_until(Wait(video, interval=1, timeout=duration),
                                   video, playback_done)
+                except TimeoutException:
+                    # We want to make sure that the current time is near the
+                    # duration we want, especially when testing partial
+                    # playback
+                    self.assertAlmostEqual(video.current_time,
+                                           start_time + duration,
+                                           delta=time_delta)
                 except VideoException as e:
                     raise self.failureException(e)
