--- firefox_media_tests/playback/video.ini
@@ -1,1 +1,0 @@
-[test_video_playback.py]
