+++ firefox_media_tests/playback/youtube/manifest.ini
@@ -1,0 +1,1 @@
+[test_basic_playback.py ]
