+++ firefox_media_tests/urls/default.ini
@@ -6,4 +6,4 @@
 # 0:08
 [https://www.youtube.com/watch?v=1visYpIREUM]
 # 2:09
-[https://www.youtube.com/watch?v=rpYRAs6ePY8]
+[https://www.youtube.com/watch?v=rjmuKV9BTkE]
